chore(dev-deps): remove lodash.set
