+++ package.json
@@ -172,7 +172,6 @@
     "jest-environment-jsdom": "^28.1.3",
     "js-yaml": "^4.1.0",
     "lockfile-lint": "^4.13.2",
-    "lodash.set": "^4.3.2",
     "mkdirp": "^3.0.1",
     "node-fetch": "^2.7.0",
     "node-mocks-http": "^1.14.1",
